Refactored super mapped classes to a different package.
