+++ VM-Core/src/main/java/com/validation/manager/core/db/mapped/VMAuditedObject.java
@@ -1,7 +1,7 @@
 /**
  * This file was generated by the Jeddict
  */
-package com.validation.manager.core.db;
+package com.validation.manager.core.db.mapped;
 
 import com.validation.manager.core.AuditedObject;
 import java.util.Date;
